refactor(api): move workflow transfer in a seperate module
